requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
asyncio==4.0.0
audioop-lts==0.2.2
botocore==1.36.20
build==1.2.2.post1
CacheControl==0.14.3
cachetools==6.2.1
certifi==2025.4.26
cffi==1.17.1
channels==4.3.1
charset-normalizer==3.4.2
cleo==2.1.0
crashtest==0.4.1
distlib==0.3.9
distro==1.9.0
Django==5.2.5
django-basicauth==0.5.3
dulwich==0.22.8
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
flake8==7.3.0
git-remote-codecommit==1.17
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.47.0
google-auth-httplib2==0.2.1
google-genai==1.59.0
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jmespath==1.0.1
keyring==25.6.0
mccabe==0.7.0
more-itertools==10.7.0
msgpack==1.1.0
mysql-connector-python==9.4.0
mysqlclient==2.2.7
packaging==25.0
pbs-installer==2025.5.17
pillow==12.0.0
pkginfo==1.12.1.2
platformdirs==4.3.8
playwright==1.55.0
poetry==2.1.3
poetry-core==2.1.3
proto-plus==1.26.1
protobuf==5.29.5
psycopg==3.2.11
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.12.3
pydantic_core==2.41.4
pyee==13.0.0
pyflakes==3.4.0
pyparsing==3.2.5
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
RapidFuzz==3.11.0
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.3
sqlparse==0.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
tenacity==9.1.2
tomlkit==0.13.2
tqdm==4.67.1
trove-classifiers==2025.1.15.22
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.3.0
virtualenv==20.31.2
websockets==15.0.1
whitenoise==6.11.0
xattr==1.1.4
zope.interface==8.0.1
zstandard==0.23.0
